140025210 add failOpen config to oauth
add failOpen and failopenGraceInterval configuration to oauth.
If failOpen is true and failopenGraceInterval is set then old token is used in case of 5xx errors in verifyApiKey call.
To use this feature set in oauth config as:

oauth:
  failopenGraceInterval: 5 (in seconds, old token will be used for this time duration)
  failOpen: true

This feature is applicable only if oauth.cacheKey is set to true.